add service directory, url service and species service classes
